Page Component View - regions are not displayed on the view, when a layout has been inserted #447
Added grid invalidation, that will render the viewport again. Reason: node update may be performed without full grid render, like when the row is added to the bottom and the `top` css property is set. This will screw up the DnD, since it depends on the Element order in the DOM.