app-contentstudio: Remove dependency from lib-admin-ui styles #586
Added CSS minification with the `cssnano`. The standalone version is used, instead of build into the `css-loader`, since the version `< 4.0.0` do not support new `browserslist` config with the `extends` keyword, that is defined in the `package.json`.
Upgraded dependencies.
Set the required node version to the latest LTS release `v8.10.0`.